+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* Enable django messages to be hidden after set timeout
 
 1.0.2 (2022-03-03)
 ==================